DeepLabv3/requirements.txt
torch
torchvision
numpy
pandas
Pillow
scikit-learn
matplotlib
segmentation_models_pytorch
transformers